+++ test/App.svelte
@@ -1,13 +1,17 @@
 <script>
-import {generateCss, parseAtoms} from "../src[vite-plugin-adorable-css]/src/atomizer"
-
-const value = `>h1:bold+mt(20)`
-
-console.log(parseAtoms(value))
-
-const css = generateCss(parseAtoms(value)).join("\n")
-
-console.log(css)
+let t = 0
+// setInterval(() => {
+//     t += 10
+// }, 1000)
 </script>
 
 <!--<Tutorial/>-->
+
+<div class="layer pack">
+  <div class="hbox gap(12)">
+    <div class="w(30) h(30) pack bg(orange)">1</div>
+    <div class="w(30) h(30) pack bg(orange)">1</div>
+    <div class="w(30) h(30) pack bg(orange)">1</div>
+    <div class="w(30) h(30) pack bg(orange)">1</div>
+  </div>
+</div>