+++ test/index.ts
@@ -1,4 +1,4 @@
-import "@adorable.css"
+// import "@adorable.css"
 import App from "./App.svelte"
 
 new App({target: document.body})